Account for iOS and OSX iMessage previews
iMessage on iOS and OSX will fetch a URL and load it up in a preview.
This broke the one time view for secrets without a password  because it
would immediately be deleted once it was fetched by iMessage.

Adding a confirmation dialogue to load up the secret whether it's
locked or not.